agentes/agente-pesquisador/requirements.txt
# agentes/agente-pesquisador/requirements.txt

flask
google-generativeai
python-dotenv
langchain-google-genai # Para os embeddings
qdrant-client